Tutorial map, pickaxe, climbing puzzle, destroyable rock and dynamic spot lights
Put in the new map, with all the assets. Added a pickaxe that can be picked up, dropped and destroy things with.
Added a destroyable rock as the first obstacle.
Added the first mini puzzle, a missing slab pops into place so you can climb up. Added lights around items of interest to give player an idea what to do.
Added a temporary skybox.